+++ functions_framework_builder/lib/src/constants.dart
@@ -12,6 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export 'src/cloud_event.dart';
-export 'src/cloud_event_wrapper.dart';
-export 'src/cloud_function.dart';
+const functionsLibraryPrefix = 'function_library';