+++ test/hello/lib/functions.dart
@@ -33,6 +33,7 @@ Future<Response> function(Request request) async {
   if (_activeRequests > _maxActiveRequests) {
     _maxActiveRequests = _activeRequests;
   }
+
   final urlPath = request.url.path;
 
   if (urlPath.contains('slow')) {
@@ -97,6 +98,41 @@ Future<Response> function(Request request) async {
   } finally {
     _activeRequests--;
   }
+}
+
+@CloudFunction()
+Future<Response> conformanceHttp(Request request) async {
+  final content = await request.readAsString();
+
+  File('function_output.json').writeAsStringSync(
+    content,
+  );
+
+  final buffer = StringBuffer()
+    ..writeln('Hello, conformance test!')
+    ..writeln('HEADERS')
+    ..writeln(_encoder.convert(request.headers))
+    ..writeln('BODY')
+    ..writeln(_encoder.convert(jsonDecode(content)));
+
+  final output = buffer.toString();
+  print(output);
+  return Response.ok(output);
+}
+
+@CloudFunction()
+void conformanceCloudEvent(CloudEvent event) {
+  final eventEncoded = _encoder.convert(event);
+  File('function_output.json').writeAsStringSync(
+    eventEncoded,
+  );
+
+  final buffer = StringBuffer()
+    ..writeln('Hello, conformance test!')
+    ..writeln('EVENT')
+    ..writeln(eventEncoded);
+
+  print(buffer.toString());
 }
 
 final _helloWorldBytes = utf8.encode('Hello, World!');